requirements.txt
streamlit
langchain
langchain_google_genai
dotenv
pandas
pypdf
python-docx